libyang example
This shows how we can depend on the libyang C library from an Acton project.
The dependency is defined in build.act.json as a zig_dependency. It refers to a
URL which contains libyang, which in this case is a branch on a fork that in
addition to the upstream libyang also has a build.zig file. Running `acton
build` to build this project will automatically download the libyang C library,
build it and link with it. All headers and paths are worked out automaticallt
and compilation flags, like optimization or which target (for
cross-compilation) automatically works correctly throughout.